bug with rsvp when not logged in

+++ course/views/classes.py
@@ -141,6 +141,9 @@ def rsvp(request,session_pk):
   session = get_object_or_404(Session,pk=session_pk)
   if session.section.fee > 0:
     raise ValueError("Some one tried to rsvp for a class that costs money!")
+  if not request.user.is_authenticated():
+    m = "You must be logged in to rsvp. Click the icon at the top right of the page to login or register"
+    return HttpResponse(simplejson.dumps([0,m,session.full]))
   enrollment,new = Enrollment.objects.get_or_create(user=request.user,session=session)
   if "drop" in request.GET:
     enrollment.delete()
